+++ frontend/src/hooks/common/useCalendar.ts
@@ -1,0 +1,56 @@
+import { useState } from 'react';
+
+import calendar from '@Utils/calendar';
+
+const useCalendar = () => {
+  const today = new Date();
+
+  const [year, setYear] = useState(today.getFullYear());
+  const [month, setMonth] = useState(today.getMonth() + 1);
+  const [navigationYear, setNavigationYear] = useState(year);
+
+  const handleMonthShift = (type: 'next' | 'prev' | 'today') => {
+    if (type === 'today') {
+      const today = new Date();
+
+      setYear(today.getFullYear());
+      setMonth(today.getMonth() + 1);
+
+      return;
+    }
+
+    const changedMonth = month + (type === 'next' ? +1 : -1);
+
+    if (changedMonth === 0) {
+      setYear((prev) => prev - 1);
+      setNavigationYear((prev) => prev - 1);
+      setMonth(12);
+      return;
+    }
+
+    if (changedMonth === 13) {
+      setYear((prev) => prev + 1);
+      setNavigationYear((prev) => prev + 1);
+      setMonth(1);
+      return;
+    }
+
+    setMonth(changedMonth);
+  };
+
+  const handleNavigationYear = (type: 'next' | 'prev') => {
+    if (type === 'next') setNavigationYear((prev) => prev + 1);
+    else setNavigationYear((prev) => prev - 1);
+  };
+
+  const handleNavigationMonth = (month: number) => {
+    setYear(navigationYear);
+    setMonth(month);
+  };
+
+  const monthStorage = calendar.getMonthStorage(year, month);
+
+  return { year, month, navigationYear, handleMonthShift, handleNavigationMonth, handleNavigationYear, monthStorage };
+};
+
+export default useCalendar;
